Move dialogue app's helper functions for creating a jsbox config into their own module
