+++ go/apps/dialogue/tests/test_vumi_app.py
@@ -14,8 +14,8 @@ from vumi.application.tests.test_sandbox import (
 from vumi.tests.helpers import VumiTestCase
 from vumi.tests.utils import LogCatcher
 
-from go.apps.dialogue.vumi_app import (
-    DialogueApplication, PollConfigResource, dialogue_js_config)
+from go.apps.dialogue.vumi_app import DialogueApplication, PollConfigResource
+from go.apps.dialogue.utils import dialogue_js_config
 from go.apps.dialogue.tests.dummy_polls import simple_poll
 from go.apps.tests.helpers import AppWorkerHelper
 
@@ -266,52 +266,3 @@ class TestPollConfigResource(ResourceTestCaseBase):
         config = json.loads(reply['value'])
 
         self.assertEqual(config, dialogue_js_config(conv))
-
-
-class TestDialogueJsConfig(VumiTestCase):
-    @inlineCallbacks
-    def setUp(self):
-        self.app_helper = self.add_helper(AppWorkerHelper(DialogueApplication))
-        self.app = yield self.app_helper.get_app_worker({})
-
-    @inlineCallbacks
-    def test_config_delivery_class(self):
-        poll = simple_poll()
-        poll['poll_metadata']['delivery_class'] = 'twitter'
-
-        conv = yield self.app_helper.create_conversation(config={'poll': poll})
-
-        config = dialogue_js_config(conv)
-        self.assertEqual(config['delivery_class'], 'twitter')
-
-    @inlineCallbacks
-    def test_config_endpoints(self):
-        poll = simple_poll()
-
-        poll['channel_types'] = [{
-            'name': 'sms',
-            'label': 'SMS'
-        }, {
-            'name': 'twitter',
-            'label': 'Twitter'
-        }]
-
-        poll['states'] = [{
-            'type': 'foo'
-        }, {
-            'type': 'send',
-            'channel_type': 'sms'
-        }, {
-            'type': 'foo'
-        }, {
-            'type': 'send',
-            'channel_type': 'twitter'
-        }, {
-            'type': 'send',
-            'channel_type': 'sms'
-        }]
-
-        conv = yield self.app_helper.create_conversation(config={'poll': poll})
-
-        config = dialogue_js_config(conv)
-        self.assertEqual(config['endpoints'], ['SMS', 'Twitter'])
